Made the data scanner a bit more fault tolerant
The previous versions would throw an exception if the content of a
dictionary was invalid b-code when parsing with the info_hash option
enabled.

To get the hash of the info dictionary we need to get the exact verbatim
bcode for the info dictionary, as we cannot rely on the order of keys to
be the same from bencoder implementation to bencoder implementation. For
this we need to get the length of the info dictionary, and that is why
there is a scanner which only purpose is to find the byte length of the
dict. This implementation was very optimistic, which would make it throw
on faulty data.

This update should fix this.